requirements.txt
Django==2.2.3
dfva-python==0.0.6
django-chunked-upload==1.1.3